% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlotly.R
\name{DimPlotly}
\alias{DimPlotly}
\title{DimPlotly}
\usage{
DimPlotly(object, grouping_var = "ident", label = FALSE,
  label_size = 12, show_arrow = FALSE, label_color = "000000",
  return = FALSE, pt_size = 4, pt_shape = "circle", opacity = 0.75,
  reduction = "tsne", dim_1 = 1, dim_2 = 2,
  palette = "default_ucscgb", plot_height = 900, plot_width = 900,
  plot_title = NULL, pt_info = NULL, legend = TRUE,
  legend_font_size = 12)
}
\arguments{
\item{object}{scRNA-seq object}

\item{grouping_var}{Variable by which to group cells. Currently only works with the current ident and column names from meta.data. Default: ident}

\item{label}{Add a label showing thr group name to the graph. Default: FALSE}

\item{label_size}{Label font size. Default: 12}

\item{show_arrow}{Offset the position of the labels and instead point to each group with an arrow. Default: FALSE}

\item{label_color}{Color for label border and arrow.  Need hex value.. Default = '000000'}

\item{return}{Return the plot object instead of displaying it. Default: FALSE}

\item{pt_size}{Size of the points in pixels. Default: 2}

\item{pt_shape}{Shape to use for the points. Default: circle}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale. Default: 1}

\item{reduction}{Dimensional reduction to display. Default: umap}

\item{dim_1}{Dimension to display on the x-axis. Default: 1}

\item{dim_2}{Dimension to display on the y-axis. Default: 2}

\item{palette}{Color palette to use.  Must be a palette available in the Paletteer package}

\item{plot_height}{Plot height in pixels. Default: 900}

\item{plot_width}{Plot width in pixels. Default: 900}

\item{plot_title}{Plot title}

\item{pt_info}{Meta.data columns to add to the hoverinfo popup.. Default: ident}

\item{legend}{Display legend?. Default: TRUE}

\item{legend_font_size}{Legend font size. Default: 12}
}
\value{
plotly object
}
\description{
Plot dimensional reduction for a Seurat object

Create a scatterplot of a given dimensional reduction set for a scRNA-seq data object,
coloring points by the given grouping_var variable
}
