% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{Feature2Plotly}
\alias{Feature2Plotly}
\title{Title}
\usage{
Feature2Plotly(seuratObj, feature.1.use = NULL, feature.2.use = NULL,
  do.return = FALSE, pt.scale = 1.5, pt.shape = "circle", alpha = 0.5,
  reduction.use = "tsne", dim.1 = 1, dim.2 = 2, colors.use.1 = "Reds",
  colors.use.2 = "Blues", bins = 10, plot.height = "750",
  plot.width = "750", pt.info = NULL, legend = TRUE)
}
\arguments{
\item{legend}{}
}
\description{
Title
}
