% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{FeaturePlotly}
\alias{FeaturePlotly}
\title{Title}
\usage{
FeaturePlotly(seuratObj, feature.use = NULL, do.return = TRUE,
  pt.scale = 5, pt.shape = "circle", alpha = 1, reduction.use = "umap",
  dim.1 = 1, dim.2 = 2, colors.use = "Reds", bins = 10,
  plot.height = "600", plot.width = "600", plot.title = FALSE,
  pt.info = NULL, legend = TRUE)
}
\arguments{
\item{legend}{}
}
\description{
Title
}
