% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{FeaturePlotly3D}
\alias{FeaturePlotly3D}
\title{Title}
\usage{
FeaturePlotly3D(seuratObj, feature.use = NULL, do.return = FALSE,
  pt.scale = 0.1, pt.shape = "circle", alpha = 1,
  reduction.use = "tsne", dim.1 = 1, dim.2 = 2, dim.3 = 3,
  colors.use = "Reds", bins = 10, plot.height = "600",
  plot.width = "600", plot.title = FALSE, pt.info = NULL, legend = TRUE,
  plot.grid = FALSE, plot.axes = FALSE)
}
\arguments{
\item{plot.axes}{}
}
\description{
Title
}
