% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_c_names}
\alias{palettes_c_names}
\title{Names of all continuous palettes}
\format{A data.frame of 205 observations with 3 variables:
\describe{
  \item{package}{character, name of package}
  \item{palette}{character, name of palette}
  \item{type}{character, type of palette}
}}
\usage{
palettes_c_names
}
\description{
Names of all continuous palettes
}
\keyword{datasets}
