% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d_names}
\alias{palettes_d_names}
\title{Names of all fixed discrete palettes}
\format{A data.frame of 869 observations with 4 variables:
\describe{
  \item{package}{character, name of package}
  \item{palette}{character, name of palette}
  \item{length}{integer, number of colors in palette}
  \item{type}{character, type of palette}
}}
\usage{
palettes_d_names
}
\description{
Names of all fixed discrete palettes
}
\keyword{datasets}
