% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{PercentAbove}
\alias{PercentAbove}
\title{PercentAbove}
\usage{
PercentAbove(x, threshold)
}
\arguments{
\item{x}{A list of numeric values.}

\item{threshold}{A numeric threshold value.}
}
\value{
double
}
\description{
Return the percentage of a list that is above a threshold
}
