% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepQuantitativePalette}
\alias{PrepQuantitativePalette}
\title{PrepQuantitativePalette}
\usage{
PrepQuantitativePalette(bins, palette)
}
\arguments{
\item{bins}{Number of discrete values that need colors.}

\item{palette}{The name of a palette to use.  Must be a palette available in
the Paletteer package.  If there are more unique identities than colors in the
palette, additional values will be created by interpolation.}
}
\value{
A list containing color values.
}
\description{
Given a palette name and data frame with clustering or grouping_var
information in an 'ident' column, return a palette containing a color for each
unique cluster identity.
}
