% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{retrieveGO}
\alias{retrieveGO}
\title{retrieveGO}
\usage{
retrieveGO(term, mart = "ensembl", dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{term}{GO ID to retreive names for.}

\item{mart}{Biomart database to use (default: 'ensembl')}

\item{dataset}{Biomart dataset to use (default: 'hsapiens_gene_ensembl')}
}
\value{
A list of HCGN names.
}
\description{
Retrieve the HUGO Gene Nomenclature Committee names associated with a GeneOntology term.
}
\examples{
genes_to_plot <- retrieveGO('GO:0046774')

}
