% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{BubblePlotly}
\alias{BubblePlotly}
\title{Title}
\usage{
BubblePlotly(object, genes.plot, colors.use = "Blues", col.min = 0,
  col.max = 10, dot.min = 0, dot.scale = 2, group.by,
  plot.legend = FALSE, do.return = FALSE, x.lab.rot = -45,
  plot.width = 600, plot.height = 600, x.font.size = 10,
  y.font.size = 10, title.font.size = 12, legend.text.size = 10,
  scale = 10, opacity = 1, plot.title = NULL, bins = 50, flip = FALSE,
  alphabetize = TRUE, pct.expr.thresh = NULL, export.df = FALSE,
  use.scaled = FALSE, use.raw = FALSE, show.zeros = FALSE,
  add.group = NULL, y.label.order = NULL)
}
\arguments{
\item{y.label.order}{}
}
\description{
Title
}
