% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubblePlotly.R
\name{BubblePlotly}
\alias{BubblePlotly}
\title{BubblePlotly}
\usage{
BubblePlotly(object, features_plot, assay = NULL, slot = "data",
  grouping_var = "ident", colors = "Blues", dot_min = 0,
  dot_scale = 2, plot_legend = FALSE, return = FALSE,
  x_label_rotation = -45, plot_width = 600, plot_height = 600,
  x_font_size = 10, y_font_size = 10, title_font_size = 12,
  legend_text_size = 10, opacity = 1, plot_title = NULL, bins = 50,
  flip = FALSE, alphabetize = TRUE, pct_expr_thres = NULL,
  show_zeros = FALSE, add_group = NULL, y_label_order = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{features_plot}{Features to display values for. If trying to retrieve values from an assay
that is not the current DefaultAssay, prefix the feature name with 'assay_'.
Works with anything \code{Seurat::\link[Seurat]{FetchData}} can retrieve.}

\item{assay}{Assay to pull data from.  Defaults to the object's DefaultAssay}

\item{slot}{Slot to pull data from.  Default: 'data'}

\item{grouping_var}{Factor by which to group cells.  Default: ident}

\item{colors}{Color palette to use.  Palettes from RColorBrewer and viridis. Default: Reds}

\item{dot_min}{Minimium marker size, in pixels. Default: 0}

\item{dot_scale}{Factor by which to scale markers. Default: 2}

\item{plot_legend}{Display legend? Default: FALSE}

\item{return}{Return the generated data frame underlying the graph. Default: FALSE}

\item{x_label_rotation}{Angle by which to rotate the x-axis labels, in degrees relative to horizontally aligned text. Default -45°}

\item{plot_width}{Plot width in pixels. Default: 900}

\item{plot_height}{Plot height in pixels. Default: 900}

\item{x_font_size}{Size of the x-axis titles. Default: 10}

\item{y_font_size}{Size of the y-axis titles. Default: 10}

\item{title_font_size}{Size of the plot title. Default: 12}

\item{legend_text_size}{Size of the legend text. Default: 10}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale Default: 1}

\item{plot_title}{Display title with the name of the feature? Default TRUE}

\item{bins}{Number of bins to use in dividing expression levels. Default: 10}

\item{flip}{Swap the x- and y-axes so that features are on the y-axis and groups along the x-axis. Default: false}

\item{alphabetize}{Alphabetize the display order of features Default: TRUE}

\item{pct_expr_thres}{Hide a feature if there is no group that expresses at or above this percentage. Default: 0}

\item{show_zeros}{Display comparison groups for which there there was no data? Default: FALSE}

\item{add_group}{Add a null comparison group.  Will be initialized with zeros.}

\item{y_label_order}{List of labels to use for the comparison groups.  Will be displayed in order of their index.}
}
\description{
Plot average expression levels and proportion expressing a feature arranged by given grouping_var.
Color intensity of the markers indicates expression levels and size of the marker indicates
the proportion of the group expressing the feature above a given threshold. By default, features are arranged
along the x-axis and groups along the y-axis.
}
