% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlotly.R
\name{DimPlotly}
\alias{DimPlotly}
\title{Plot dimensional reduction for a Seurat object}
\usage{
DimPlotly(object, grouping_var = "ident", do.label = FALSE,
  label.size = 12, show.arrow = FALSE, label.color = "000000",
  return = FALSE, pt.size = 4, pt_shape = "circle", opacity = 0.75,
  reduction = "tsne", dim_1 = 1, dim_2 = 2,
  palette = "default_ucscgb", plot_height = 900, plot_width = 900,
  plot_title = NULL, pt_info = NULL, legend = TRUE,
  legend_font_size = 12)
}
\arguments{
\item{object}{scRNA-seq object}

\item{grouping_var}{Variable by which to group cells. Currently only works with the current ident and column names from meta.data (default: ident)}

\item{do.label}{Add a label showing thr group name to the graph (default: FALSE)}

\item{label.size}{Label font size (default: 12)}

\item{show.arrow}{Offset the position of the labels and instead point to each group with an arrow (default: FALSE)}

\item{label.color}{Color for label border and arrow.  Need hex value. (default = '000000')}

\item{return}{Return the plot object instead of displaying it (default: FALSE)}

\item{pt.size}{Size of the points in pixels (default: 2)}

\item{pt_shape}{Shape to use for the points (default: circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale (default: 1)}

\item{reduction}{Dimensional reduction to display (default: tsne)}

\item{dim_1}{Dimension to display on the x-axis (default: 1)}

\item{dim_2}{Dimension to display on the y-axis (default: 2)}

\item{palette}{Color palette to use.  Must be a palette available in the Paletteer package}

\item{plot_height}{Plot height in pixels (default: 900)}

\item{plot_width}{Plot width in pixels (default: 900)}

\item{pt_info}{Meta.data columns to add to the hoverinfo popup. (default: ident)}

\item{legend}{Display legend? (default: TRUE)}

\item{legend_font_size}{Legend font size (default: 12)}
}
\value{
plotly object
}
\description{
Create a scatterplot of a given dimensional reduction set for a scRNA-seq data object,
coloring points by the given grouping_var variable
}
\examples{
object <- RunTSNE(object)
DimPlotly(object, grouping_var = 'ident', pt.size = 4, opacity = 0.5, plot_title = "Test Plot", reduction = "tsne")
}
