% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{DimPlotly}
\alias{DimPlotly}
\title{Plot dimensional reduction for a Seurat object}
\usage{
DimPlotly(seuratObj, group.by = "ident", do.label = FALSE,
  label.size = 12, show.arrow = FALSE, do.return = FALSE, pt.size = 2,
  pt.shape = "circle", opacity = 1, reduction.use = "umap", dim.1 = 1,
  dim.2 = 2, colors.use = "Set1", plot.height = "600",
  plot.width = "600", pt.info = NULL, legend = TRUE)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{group.by}{Variable by which to group cells. Currently only works with the current ident and column names from meta.data (default: ident)}

\item{do.label}{Add a label showing thr group name to the graph (default: FALSE)}

\item{label.size}{Label font size (default: 12)}

\item{show.arrow}{Offset the position of the labels and instead point to each group with an arrow (default: FALSE)}

\item{do.return}{Return the plot object instead of displaying it (default: FALSE)}

\item{pt.size}{Size of the points in pixels (default: 2)}

\item{pt.shape}{Shape to use for the points (default: circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale (default: 1)}

\item{reduction.use}{Dimensional reduction to display (default: umap)}

\item{dim.1}{Dimension to display on the x-axis (default: 1)}

\item{dim.2}{Dimension to display on the y-axis (default: 2)}

\item{colors.use}{Color palette to use.  Palettes from RColorBrewer and viridis}

\item{plot.height}{Plot height in pixels (default: 600)}

\item{plot.width}{Plot width in pixels (default: 600)}

\item{pt.info}{Meta.data columns to add to the hoverinfo popup. (default: ident)}

\item{legend}{Display legend (default TRUE)}
}
\value{
plotly object
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring points by the given grouping variable
Requrires a Seurat object with the reduction to be used in the corresponding
seuratObj@dr slot
}
\examples{
seuratObj <- RunTSNE(seuratObj)
DimPlotly(seuratObj, group.by = 'ident', pt.size = 4, opacity = 0.5, plot.title = "Test Plot", reduction.use = "tsne")
}
