% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlotly3D.R
\name{DimPlotly3d}
\alias{DimPlotly3d}
\title{DimPlotly3D}
\usage{
DimPlotly3d(object, grouping_var = "ident", label = FALSE,
  label_size = 12, label_color = "000000", show_arrow = FALSE,
  return = FALSE, pt_size = 2, pt_shape = "circle", opacity = 1,
  reduction = "tsne", dim_1 = 1, dim_2 = 2, dim_3 = 3,
  palette = "Set1", plot_height = 900, plot_width = 900,
  plot_title = NULL, pt_info = NULL, legend = TRUE,
  legend_font_size = 12, plot_grid = FALSE, plot_axes = FALSE)
}
\arguments{
\item{object}{Seurat object}

<<<<<<< HEAD
\item{grouping_var}{Variable by which to group cells. Currently only works with the current ident and column names from meta.data (default: ident)}
=======
\item{grouping}{Variable by which to group cells. Currently only works with the current ident and column names from meta.data. Default: ident}
>>>>>>> 6f6c82c504d67862c3ae3c061af7383aa47a7f7f

\item{label}{Add a label showing thr group name to the graph. Default: FALSE}

\item{label_size}{Label font size. Default: 12}

\item{label_color}{Color for label border and arrow.  Need hex value.. Default = '000000'}

\item{show_arrow}{Offset the position of the labels and instead point to each group with an arrow. Default: FALSE}

\item{return}{Return the plot object instead of displaying it. Default: FALSE}

\item{pt_size}{Size of the points in pixels. Default: 2}

\item{pt_shape}{Shape to use for the points. Default: circle}

\item{opacity}{Transparency level to use for the points on a 0-1 scale. Default: 1}

<<<<<<< HEAD
\item{reduction}{Dimensional reduction to display (default: tsne)}
=======
\item{reduction_use}{Dimensional reduction to display. Default: tsne}
>>>>>>> 6f6c82c504d67862c3ae3c061af7383aa47a7f7f

\item{dim_1}{Dimension to display on the x-axis. Default: 1}

\item{dim_2}{Dimension to display on the y-axis. Default: 2}

\item{dim_3}{Dimension to display on the z-axis. Default: 3}

<<<<<<< HEAD
\item{palette}{Color palette to use.  Must be a palette available in the Paletteer package.  (default: 'Set1')}
=======
\item{palette_use}{Color palette to use.  Must be a palette available in the Paletteer package. . Default: 'Set1'}
>>>>>>> 6f6c82c504d67862c3ae3c061af7383aa47a7f7f

\item{plot_height}{Plot height in pixels. Default: 900}

\item{plot_width}{Plot width in pixels. Default: 900}

<<<<<<< HEAD
\item{plot_title}{Plot title (default: reduction)}
=======
\item{plot_title}{Plot title. Default: reduction_use}
>>>>>>> 6f6c82c504d67862c3ae3c061af7383aa47a7f7f

\item{pt_info}{Meta.data columns to add to the hoverinfo popup.. Default: ident}

\item{legend}{Display legend?. Default: TRUE}

\item{legend_font_size}{Legend font size. Default: 12}

\item{plot_grid}{Display the major unit tick marks?. Default: FALSE}

\item{plot_axes}{Display the major x, y, and z axes?. Default: FALSE}
}
\value{
plotly object
}
\description{
DimPlotly3D
}
\examples{
DimPlotly3D(object, grouping_var = "res.0.6", label = TRUE, show_arrow = FALSE)
}
