% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature2Plotly.R
\name{Feature2Plotly}
\alias{Feature2Plotly}
\title{Feature2Plotly}
\usage{
Feature2Plotly(object, feature_1 = NULL, feature_2 = NULL,
  assay_1 = "RNA", assay_2 = "RNA", return = FALSE, pt_scale = 1.5,
  pt_shape = "circle", opacity = 0.5, reduction_use = "tsne",
  dim_1 = 1, dim_2 = 2, colors_use_1 = "Reds",
  colors_use_2 = "Blues", bins = 10, plot_height = "750",
  plot_width = "750", pt_info = NULL, legend = TRUE,
  legend_font_size = 12)
}
\arguments{
\item{object}{Seurat object}

\item{feature_1}{First variable to display. Currently only works with gene names}

\item{feature_2}{Second variable to display. Currently only works with gene names}

\item{return}{Return the plot dataframe instead of displaying it Default: FALSE)}

\item{pt_scale}{Factor by which to multiply the size of the points Default: 5)}

\item{pt_shape}{Shape to use for the points Default = circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale Default: 1)}

\item{reduction_use}{Dimensional reduction to display Default: tsne)}

\item{dim_1}{Dimension to display on the x-axis Default: 1)}

\item{dim_2}{Dimension to display on the y-axis Default: 2)}

\item{colors_use_1}{Color palette to use for feature 1.  Palettes from RColorBrewer and viridis. Default: Reds)}

\item{colors_use_2}{Color palette to use for feature 2.  Palettes from RColorBrewer and viridis. Default: Reds)}

\item{bins}{Number of bins to use in dividing expression levels. Default: 10)}

\item{plot_height}{Plot height in pixels Default: 900)}

\item{plot_width}{Plot width in pixels Default: 900)}

\item{pt_info}{Meta.data columns to add to the hoverinfo popup. Default: ident)}

\item{legend}{Display legend? Default: TRUE)}

\item{legend_font_size}{Legend font size Default: 12)}

\item{plot_title}{Display title with the name of the feature? Default TRUE)}
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring and sizing points by the expression level of two different features.
}
