% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature2Plotly3D.R
\name{Feature2Plotly3D}
\alias{Feature2Plotly3D}
\title{Feature2Plotly3D}
\usage{
Feature2Plotly3D(object, reduction = "umap", feature_1 = NULL,
  feature_2 = NULL, colors_1 = "Reds", colors_2 = "Blues",
  dim_1 = 1, dim_2 = 2, dim_3 = 3, bins = 10, return = FALSE,
  pt_scale = 0.5, pt_shape = "circle", opacity = 1,
  plot_height = "750", plot_width = "750", pt_info = NULL,
  legend = TRUE, legend_font_size = 12, plot_grid = FALSE,
  plot_axes = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction to display. Default: 'umap'}

\item{feature_1}{First feature values to display. Works with anything \code{Seurat::\link[Seurat]{FetchData}} can retrieve.}

\item{feature_2}{Second feature values to display.}

\item{colors_1}{Colors to use to display values for feature_1. Default: "Reds"}

\item{colors_2}{Colors to use to display values for feature_2. Default: "Blues"}

\item{dim_1}{Dimension to display on the x-axis. Default: 1}

\item{dim_2}{Dimension to display on the y-axis. Default: 2}

\item{dim_3}{Dimension to display on the z-axis. Default: 3}

\item{bins}{Number of bins to use in dividing expression levels.. Default: 10}

\item{return}{Return the plot object instead of displaying it. Default: FALSE}

\item{pt_scale}{Factor by which to multiply the size of the points. Default: 5}

\item{pt_shape}{Shape to use for the points. Default = circle}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale. Default: 1}

\item{plot_height}{Plot height in pixels. Default: 900}

\item{plot_width}{Plot width in pixels. Default: 900}

\item{pt_info}{Meta.data columns to add to the hoverinfo popup.. Default: ident}

\item{legend}{Display legend?. Default: TRUE}

\item{legend_font_size}{Legend font size. Default: 12}

\item{plot_grid}{Display the major unit tick marks?. Default: FALSE}

\item{plot_axes}{Display the major x, y, and z axes?. Default: FALSE}

\item{assay_1}{Assay to pull values from for feature_1. Default: NULL}

\item{assay_2}{Assay to pull values from for feature_2. Default: NULL}
}
\value{
If return is TRUE, a plotly object.
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring and sizing points by the expression level of the chosen feature.
Requrires a Seurat object with the reduction to be used in the corresponding
object@dr slot
}
