% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature2Plotly3D.R
\name{Feature2Plotly3D}
\alias{Feature2Plotly3D}
\title{Feature2Plotly3D}
\usage{
Feature2Plotly3D(seuratObj, feature.1.use = NULL, feature.2.use = NULL,
  do.return = FALSE, pt.scale = 0.5, pt.shape = "circle",
  opacity = 1, reduction.use = "tsne", dim.1 = 1, dim.2 = 2,
  dim.3 = 3, colors.use.1 = "Reds", colors.use.2 = "Blues",
  bins = 10, plot.height = "750", plot.width = "750",
  pt.info = NULL, legend = TRUE, legend.font.size = 12,
  plot.grid = FALSE, plot.axes = FALSE)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{feature.1.use}{First variable to display. Currently only works with gene names}

\item{feature.2.use}{Second variable to display. Currently only works with gene names}

\item{do.return}{Return the plot object instead of displaying it (default: FALSE)}

\item{pt.scale}{Factor by which to multiply the size of the points (default: 5)}

\item{pt.shape}{Shape to use for the points (default = circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale (default: 1)}

\item{reduction.use}{Dimensional reduction to display (default: tsne)}

\item{dim.1}{Dimension to display on the x-axis (default: 1)}

\item{dim.2}{Dimension to display on the y-axis (default: 2)}

\item{dim.3}{Dimension to display on the z-axis (default: 3)}

\item{colors.use.1}{Color palette to use for feature 1.  Palettes from RColorBrewer and viridis. (default: Reds)}

\item{colors.use.2}{Color palette to use for feature 2.  Palettes from RColorBrewer and viridis. (default: Reds)}

\item{bins}{Number of bins to use in dividing expression levels. (default: 10)}

\item{plot.height}{Plot height in pixels (default: 900)}

\item{plot.width}{Plot width in pixels (default: 900)}

\item{pt.info}{Meta.data columns to add to the hoverinfo popup. (default: ident)}

\item{legend}{Display legend? (default: TRUE)}

\item{legend.font.size}{Legend font size (default: 12)}

\item{plot.grid}{Display the major unit tick marks? (default: FALSE)}

\item{plot.axes}{Display the major x, y, and z axes? (default: FALSE)}

\item{plot.title}{Display title with the name of the feature? (default TRUE)}
}
\value{
If do.return is TRUE, a plotly object.
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring and sizing points by the expression level of the chosen feature.
Requrires a Seurat object with the reduction to be used in the corresponding
seuratObj@dr slot
}
