% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{FeaturePlotly}
\alias{FeaturePlotly}
\title{FeaturePlotly}
\usage{
FeaturePlotly(seuratObj, feature.use = NULL, do.return = TRUE,
  pt.scale = 5, pt.shape = "circle", opacity = 1,
  reduction.use = "umap", dim.1 = 1, dim.2 = 2, colors.use = "Reds",
  bins = 10, plot.height = "600", plot.width = "600",
  plot.title = FALSE, pt.info = NULL, legend = TRUE)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{feature.use}{Variable to display. Currently only works with gene names}

\item{do.return}{Return the plot object instead of displaying it (default: FALSE)}

\item{pt.scale}{Factor by which to multiply the size of the points (default: 5)}

\item{pt.shape}{Shape to use for the points (default = circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale (default: 1)}

\item{reduction.use}{Dimensional reduction to display (default: umap)}

\item{dim.1}{Dimension to display on the x-axis (default: 1)}

\item{dim.2}{Dimension to display on the y-axis (default: 2)}

\item{colors.use}{Color palette to use.  Palettes from RColorBrewer and viridis. (default: Reds)}

\item{bins}{Number of bins to use in dividing expression levels. (default: 10)}

\item{plot.height}{Plot height in pixels (default: 600)}

\item{plot.width}{Plot width in pixels (default: 600)}

\item{plot.title}{Display title with the name of the feature? (default TRUE)}

\item{pt.info}{Meta.data columns to add to the hoverinfo popup. (default: ident)}

\item{legend}{Display legend (default TRUE)}
}
\value{
plotly object
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring and sizing points by the expression level of the chosen feature.
Requrires a Seurat object with the reduction to be used in the corresponding
seuratObj@dr slot
}
