% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePlotly3D.R
\name{FeaturePlotly3D}
\alias{FeaturePlotly3D}
\title{FeaturePlotly3D}
\usage{
FeaturePlotly3D(object, feature_use = NULL, return = FALSE,
  pt_scale = 0.5, pt_shape = "circle", opacity = 1,
  reduction_use = "tsne", dim_1 = 1, dim_2 = 2, dim_3 = 3,
  colors_use = "Reds", bins = 10, plot_height = 900,
  plot_width = 900, plot_title = FALSE, pt_info = NULL,
  legend = TRUE, legend_font_size = 12, plot_grid = FALSE,
  plot_axes = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{feature_use}{Variable to display. Currently only works with gene names}

\item{return}{Return the plot object instead of displaying it (default: FALSE)}

\item{pt_scale}{Factor by which to multiply the size of the points (default: 5)}

\item{pt_shape}{Shape to use for the points (default = circle)}

\item{opacity}{Transparency level to use for the points, on a 0-1 scale (default: 1)}

\item{reduction_use}{Dimensional reduction to display (default: tsne)}

\item{dim_1}{Dimension to display on the x-axis (default: 1)}

\item{dim_2}{Dimension to display on the y-axis (default: 2)}

\item{dim_3}{Dimension to display on the z-axis (default: 3)}

\item{colors_use}{Color palette to use.  Palettes from RColorBrewer and viridis. (default: Reds)}

\item{bins}{Number of bins to use in dividing expression levels. (default: 10)}

\item{plot_height}{Plot height in pixels (default: 900)}

\item{plot_width}{Plot width in pixels (default: 900)}

\item{plot_title}{Display title with the name of the feature? (default TRUE)}

\item{pt_info}{Meta.data columns to add to the hoverinfo popup. (default: ident)}

\item{legend}{Display legend? (default: TRUE)}

\item{legend_font_size}{Legend font size (default: 12)}

\item{plot_grid}{Display the major unit tick marks? (default: FALSE)}

\item{plot_axes}{Display the major x, y, and z axes? (default: FALSE)}
}
\value{
If return is TRUE, a plotly object.
}
\description{
Create a scatterplot of a given dimensional reduction set for a Seurat object,
coloring and sizing points by the expression level of the chosen feature.
Requrires a Seurat object with the reduction to be used in the corresponding
object@dr slot
}
