% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{GetFeatureValues}
\alias{GetFeatureValues}
\title{GetFeatureValues}
\usage{
GetFeatureValues(object, ...)
}
\arguments{
\item{object}{Seurat object to get feature data from}

\item{df}{Plotting data frame to which to add the popup data.}

\item{feature}{feature (genes expression, metadata, etc...) to retrieve data for}

\item{assay_use}{Assay to pull feature data from.  Default: "RNA"}

\item{slot_use}{Slot to pull feature data from. Default: "data"}
}
\value{
data.frame
}
\description{
Get expression levels of a given feature for each cell
}
