% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{GetFeatureValues}
\alias{GetFeatureValues}
\alias{GetFeatureValues.Seurat}
\title{GetFeatureValues}
\usage{
GetFeatureValues(object, ...)

\method{GetFeatureValues}{Seurat}(object, features, df = NULL,
  assay = NULL, slot = NULL, bins = NULL, suffix = NULL, ...)
}
\arguments{
\item{object}{Seurat object to get feature data from}

\item{df}{Plotting data frame to which to add the popup data.}

\item{assay}{Assay to pull feature data from.  Default: NULL}

\item{slot}{Slot to pull feature data from. Default: NULL}

\item{feature}{Feature (genes expression, metadata, etc...) to retrieve data for}
}
\value{
data.frame

data.frame
}
\description{
Get expression levels of a given feature for each cell
}
