% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{GetFeatureValues}
\alias{GetFeatureValues}
\alias{GetFeatureValues.seurat}
\alias{GetFeatureValues.Seurat}
\title{GetFeatureValues}
\usage{
GetFeatureValues(object, ...)

\method{GetFeatureValues}{seurat}(object, df, feature, bins = NULL, ...)

\method{GetFeatureValues}{Seurat}(object, features, df = NULL,
  assay_use = NULL, slot_use = NULL, bins = NULL, suffix = NULL,
  ...)
}
\arguments{
\item{object}{Seurat object to get feature data from}

\item{df}{Plotting data frame to which to add the popup data.}

\item{feature}{feature (genes expression, metadata, etc...) to retrieve data for}

\item{assay_use}{Assay to pull feature data from.  Default: "RNA"}

\item{slot_use}{Slot to pull feature data from. Default: "data"}
}
\value{
data.frame

data.frame

data.frame
}
\description{
Get expression levels of a given feature for each cell
}
