% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratPlotly.R
\name{GOBubblePlotly}
\alias{GOBubblePlotly}
\title{Title}
\usage{
GOBubblePlotly(dataset, go_term, group.by = "ident", plot.height = 900,
  plot.width = 900, filter = dataset@var.genes, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
