% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOBubblePlotly.R
\name{GOBubblePlotly}
\alias{GOBubblePlotly}
\title{GOBubblePlotly}
\usage{
GOBubblePlotly(object, go_term, grouping_var = "ident",
  plot_height = 900, plot_width = 900, gene_filter = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{go_term}{Gene Ontology term identifier (i.e. GO:0046774)}

\item{grouping_var}{Factor by which to group cells.  (default: ident)}

\item{plot_height}{Plot height in pixels. (default: 900)}

\item{plot_width}{Plot width in pixels. (default: 900)}
}
\description{
Produces a Bubble Plot for the genes of a given GO term.
}
