% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepDf}
\alias{PrepDf}
\title{PrepDf}
\usage{
PrepDf(seuratObj, reduction.use, dim.1 = 1, dim.2 = 2, dim.3 = NULL,
  group.by)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{reduction.use}{dimensional reduction to extract}

\item{dim.1}{Dimension one.  X-axis data.}

\item{dim.2}{Dimension two.  Y-axis data.}

\item{dim.3}{Dimension three.  Z-axis data.}

\item{group.by}{Identity information to use for grouping.
Permissible values include meta.data column names.}
}
\value{
data.frame
}
\description{
Helper function to extract dimensional reduction data from a Seurat object
and format it into a data frame for use in plotting.
}
