% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepDf}
\alias{PrepDf}
\alias{PrepDf.seurat}
\alias{PrepDf.Seurat}
\title{PrepDf}
\usage{
PrepDf(object, ...)

\method{PrepDf}{seurat}(object, reduction, dim_1 = 1, dim_2 = 2,
  dim_3 = NULL, grouping_var = NULL)

\method{PrepDf}{Seurat}(object, reduction, dim_1 = 1, dim_2 = 2,
  dim_3 = NULL, grouping_var = NULL)
}
\arguments{
\item{object}{scRNA-seq data object}

\item{reduction}{dimensional reduction to extract}

\item{dim_1}{Dimension one.  X-axis data.}

\item{dim_2}{Dimension two.  Y-axis data.}

\item{dim_3}{Dimension three.  Z-axis data.}

\item{grouping_var}{Identity information to use for grouping_var.
Permissible values include meta.data column names.}
}
\value{
data.frame

data.frame
}
\description{
Helper function to extract dimensional reduction data from a Seurat object
and format it into a data frame for use in plotting.
}
