% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepInfo}
\alias{PrepInfo}
\title{PrepInfo}
\usage{
PrepInfo(seuratObj, pt.info, df)
}
\arguments{
\item{seuratObj}{Seurat object to extract data from}

\item{pt.info}{A list of meta.data columns to add to the hoverinfo popup.}

\item{df}{Plotting data frame to which to add the popup data.}
}
\value{
data.frame
}
\description{
Using the members of pt.info, build a string for
each row that compiles the indicated pt.info from the meta.data
slot along with the necessary html tags and column names
}
