% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepInfo}
\alias{PrepInfo}
\alias{PrepInfo.seurat}
\alias{PrepInfo.Seurat}
\alias{GetFeatureValues.seurat}
\alias{GetFeatureValues.Seurat}
\title{PrepInfo}
\usage{
PrepInfo(object, ...)

\method{PrepInfo}{seurat}(object, pt_info, df)

\method{PrepInfo}{Seurat}(object, pt_info, df)

\method{PrepInfo}{seurat}(object, df, feature, bins = NULL, ...)

\method{PrepInfo}{Seurat}(object, df = NULL, features,
  assay_use = "RNA", slot_use = "data", bins = NULL, suffix = NULL,
  ...)
}
\arguments{
\item{object}{Seurat object to extract data from}

\item{pt_info}{A list of meta.data columns to add to the hoverinfo popup.}

\item{df}{Plotting data frame to which to add the popup data.}
}
\value{
data.frame

data.frame

data.frame

data.frame

data.frame
}
\description{
Using the members of pt_info, build a string for
each row that compiles the indicated pt_info from the meta.data
slot along with the necessary html tags and column names
}
