% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{PrepInfo}
\alias{PrepInfo}
\alias{PrepInfo.Seurat}
\title{PrepInfo}
\usage{
PrepInfo(object, ...)

\method{PrepInfo}{Seurat}(object, pt_info, df)
}
\arguments{
\item{object}{Seurat object to extract data from}

\item{...}{Unused}

\item{pt_info}{A list of meta.data columns to add to the hoverinfo popup.}

\item{df}{Plotting data frame to which to add the popup data.}
}
\value{
data.frame

data.frame
}
\description{
Using the members of pt_info, build a string for
each row that compiles the indicated pt_info from the meta.data
slot along with the necessary html tags and column names
}
